# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CalendarSync OS API
  description: 複数のGoogleカレンダーを双方向同期するシステムのAPI仕様
  version: 1.0.0
  contact:
    name: CalendarSync OS Support
servers:
  - url: http://localhost:3000/api
    description: 開発環境
  - url: https://api.calendarsync.example.com/api
    description: 本番環境
...
paths:
...
  /sync/manual:
    post:
      tags:
        - Sync
      summary: 手動同期実行
      security:
        - sessionAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                calendarIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: 同期開始
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  calendarsQueued:
                    type: number
                  jobIds:
                    type: array
                    items:
                      type: string
  /sync/logs:
    get:
      tags:
        - Sync
      summary: 同期ログ取得
      security:
        - sessionAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: 同期ログ
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/SyncLog'
  /sync/history:
    get:
      tags:
        - Sync
      summary: 同期履歴取得
      security:
        - sessionAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 同期履歴
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SyncLog'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
  /sync/status:
    get:
      tags:
        - Sync
      summary: 同期ステータス取得（全体統計）
      security:
        - sessionAuth: []
      responses:
        '200':
          description: 同期ステータス
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabledCalendars:
                    type: number
                  totalCalendars:
                    type: number
                  successRate:
                    type: number
                  errorCount:
                    type: number
                  last7Days:
                    type: object
                    properties:
                      success:
                        type: number
                      errors:
                        type: number
                      total:
                        type: number
  /sync/status/{jobId}:
    get:
      tags:
        - Sync
      summary: 同期ステータス取得（ジョブID指定）
      security:
        - sessionAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ジョブステータス
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                  state:
                    type: string
                    enum: [completed, active, delayed, failed, waiting]
                  progress:
                    type: number
                  result:
                    type: object
                  failedReason:
                    type: string
  /conflicts:
    get:
      tags:
        - Conflicts
      summary: 競合一覧取得
      security:
        - sessionAuth: []
      responses:
        '200':
          description: 競合一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  conflicts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conflict'
  /conflicts/{id}:
    get:
      tags:
        - Conflicts
      summary: 競合詳細取得
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 競合詳細
          content:
            application/json:
              schema:
                type: object
                properties:
                  conflict:
                    $ref: '#/components/schemas/Conflict'
        '404':
          description: 競合が見つかりません
  /conflicts/{id}/resolve:
    post:
      tags:
        - Conflicts
      summary: 競合解決
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - strategy
              properties:
                strategy:
                  type: string
                  enum: [adopt-A, adopt-B, manual]
                adoptLinkId:
                  type: string
                  format: uuid
                manualData:
                  type: object
                  properties:
                    title:
                      type: string
                    start_at:
                      type: string
                      format: date-time
                    end_at:
                      type: string
                      format: date-time
                    location:
                      type: string
                    description:
                      type: string
      responses:
        '200':
          description: 解決成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  canonicalId:
                    type: string
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: セッション認証
  schemas:
...
    Conflict:
      type: object
      properties:
        canonicalId:
          type: string
          format: uuid
        title:
          type: string
        variants:
          type: array
          items:
            type: object
            properties:
              eventLinkId:
                type: string
                format: uuid
              accountEmail:
                type: string
                format: email
              calendarName:
                type: string
              data:
                type: object
                properties:
                  start:
                    type: string
                    format: date-time
                  end:
                    type: string
                    format: date-time
                  location:
                    type: string
                    nullable: true
                  description:
                    type: string
                    nullable: true
              lastModified:
                type: string
                format: date-time
    SyncLog:
      type: object
      properties:
        id:
          type: integer
        timestamp:
          type: string
          format: date-time
        operation:
          type: string
        from_account_id:
          type: string
          format: uuid
        to_account_id:
          type: string
          format: uuid
        event_id:
          type: string
        result:
          type: string
        error:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
...